refactor docker utils
