+++ bentoml/yatai/deployment/docker_utils.py
@@ -1,0 +1,17 @@
+import subprocess
+
+from bentoml.exceptions import BentoMLException, MissingDependencyException
+
+
+def ensure_docker_available_or_raise():
+    try:
+        subprocess.check_output(['docker', 'info'])
+    except subprocess.CalledProcessError as error:
+        raise BentoMLException(
+            'Error executing docker command: {}'.format(error.output.decode())
+        )
+    except FileNotFoundError:
+        raise MissingDependencyException(
+            'Docker is required for this deployment. Please visit '
+            'www.docker.com for instructions'
+        )
